+++ src/main/java/org/example/battleunits/Vampire.java
@@ -1,12 +1,13 @@
 package org.example.battleunits;
 
+import org.example.battleunits.characteristics.Health;
+import org.example.battleunits.subsidiary.DealtDamageAwareness;
 import org.example.battleunits.units.VampireUnit;
 import org.example.battleunits.units.WarriorUnit;
 import org.jetbrains.annotations.NotNull;
 
-public class Vampire extends Warrior implements VampireUnit {
+public class Vampire extends Warrior implements VampireUnit, DealtDamageAwareness {
     private final int vampirism;
-    private final int PERCENTS = 100;
 
     /**
      * Constructs default Vampire object with default health(40), attack(4) & vampirism(50).
@@ -31,10 +32,8 @@ public class Vampire extends Warrior implements VampireUnit {
 
     @Override
     public void hit(WarriorUnit opponent) {
-        int healthBeforeHit = opponent.getHealth();
-        super.hit(opponent);
-        int healthAfterHit = opponent.getHealth();
-        int damageDealt = Math.min(healthBeforeHit, healthBeforeHit - healthAfterHit);
+        int damageDealt = getDealtDamage(opponent);
+        final int PERCENTS = 100;
         int healingPoints = damageDealt * getVampirism() / PERCENTS;
         vampirism(healingPoints);
     }
