+++ src/main/java/org/example/battleunits/subsidiary/HealChainUnit.java
@@ -1,0 +1,10 @@
+package org.example.battleunits.subsidiary;
+
+import org.example.battleunits.units.WarriorUnit;
+
+public interface HealChainUnit {
+
+    void setNextChain(HealChainUnit nextChain);
+
+    void heal(WarriorUnit unitInFront);
+}
