+++ src/main/java/org/example/battleunits/subsidiary/WarriorUnitBehind.java
@@ -1,0 +1,8 @@
+package org.example.battleunits.subsidiary;
+
+import org.example.battleunits.units.WarriorUnit;
+
+public interface WarriorUnitBehind {
+
+    WarriorUnit getWarriorBehind();
+}
