--- src/main/java/org/example/battleunits/units/WarriorUnitBehind.java
@@ -1,6 +1,0 @@
-package org.example.battleunits.units;
-
-public interface WarriorUnitBehind {
-
-    WarriorUnit getWarriorBehind();
-}
